app/requirements.txt
docker-py
selenium
bottle
gevent
uwsgi
jinja2
